+++ fluent_blogs/admin/__init__.py
@@ -1,5 +1,5 @@
 from django.contrib import admin
-from fluent_blogs.admin.entryadmin import AbstractEntryBaseAdminForm, AbstractTranslatableEntryBaseAdminForm, AbstractEntryBaseAdmin, AbstractTranslatableEntryBaseAdmin, EntryAdmin
+from fluent_blogs.admin.entryadmin import AbstractEntryBaseAdminForm, AbstractTranslatableEntryBaseAdminForm, AbstractEntryBaseAdmin, AbstractTranslatableEntryBaseAdmin, SeoEntryAdminMixin, EntryAdmin
 from fluent_blogs.models import get_entry_model, Entry
 
 __all__ = (
@@ -8,6 +8,7 @@ __all__ = (
 
     'AbstractEntryBaseAdmin',
     'AbstractTranslatableEntryBaseAdmin',
+    'SeoEntryAdminMixin',
     'EntryAdmin',
 )